+++ CHANGELOG.md
@@ -6,6 +6,9 @@ __Feature__:
 - allow full export for tables and use partition column only to speed up extraction
 - allow to force full export. Useful for re-init cases.
 - allow to change date and timestamp format during data extraction
+- allow to prevent extraction if last extraction is recent enough
+- allow to clean all files of table only when it is extracted
+- allow to in/exclude schemas and/or tables from extraction
 - allow to filter table data before sinking it.
 
 __Bug Fix__: